# Minor optimisation to require/provide setup