--- a/src/core/build_target_benchmark.go
+++ b/src/core/build_target_benchmark.go
@@ -16,7 +16,7 @@ func BenchmarkProvideFor(b *testing.B) {
 	b.Run("Simple", func(b *testing.B) {
 		b.ReportAllocs()
 		for i := 0; i < b.N; i++ {
-			result = target1.ProvideFor(target2)
+			result = target1.provideFor(target2)
 		}
 	})
 
@@ -25,7 +25,7 @@ func BenchmarkProvideFor(b *testing.B) {
 	b.Run("NoMatch", func(b *testing.B) {
 		b.ReportAllocs()
 		for i := 0; i < b.N; i++ {
-			result = target2.ProvideFor(target1)
+			result = target2.provideFor(target1)
 		}
 	})
 
@@ -34,7 +34,7 @@ func BenchmarkProvideFor(b *testing.B) {
 	b.Run("OneMatch", func(b *testing.B) {
 		b.ReportAllocs()
 		for i := 0; i < b.N; i++ {
-			result = target2.ProvideFor(target1)
+			result = target2.provideFor(target1)
 		}
 	})
 
@@ -42,7 +42,7 @@ func BenchmarkProvideFor(b *testing.B) {
 	b.Run("TwoMatches", func(b *testing.B) {
 		b.ReportAllocs()
 		for i := 0; i < b.N; i++ {
-			result = target2.ProvideFor(target1)
+			result = target2.provideFor(target1)
 		}
 	})
 
@@ -50,7 +50,7 @@ func BenchmarkProvideFor(b *testing.B) {
 	b.Run("IsData", func(b *testing.B) {
 		b.ReportAllocs()
 		for i := 0; i < b.N; i++ {
-			result = target2.ProvideFor(target1)
+			result = target2.provideFor(target1)
 		}
 	})
 }