Tries to reduce the number of allocations by avoiding creating additional slices unless necessary.