@peterebden 2021-06-17
OK yup now it is
```
BenchmarkProvideFor
BenchmarkProvideFor/Simple
BenchmarkProvideFor/Simple-12           63322402                18.81 ns/op            0 B/op          0 allocs/op
BenchmarkProvideFor/NoMatch
BenchmarkProvideFor/NoMatch-12          31253348                38.50 ns/op            0 B/op          0 allocs/op
BenchmarkProvideFor/OneMatch
BenchmarkProvideFor/OneMatch-12          8090606               145.7 ns/op            48 B/op          1 allocs/op
BenchmarkProvideFor/TwoMatches
BenchmarkProvideFor/TwoMatches-12        4717500               247.8 ns/op            96 B/op          1 allocs/op
BenchmarkProvideFor/IsData
BenchmarkProvideFor/IsData-12            8401734               128.9 ns/op            48 B/op          1 allocs/op
```
Slice allocation is quite slow initially; looks like it allocates 1, 2, 4, 8... so preallocating helps for the case where we have > 1 match.