@tiagovtristao 2021-06-17
> OK, I have simplified out some of the nonsense (that made more sense when I wrote it...) and done what I should have originally: write a benchmark!
> 
> Before:
> 
> ```
> BenchmarkProvideFor
> BenchmarkProvideFor/Simple
> BenchmarkProvideFor/Simple-12           10824295               112.7 ns/op            48 B/op          1 allocs/op
> BenchmarkProvideFor/NoMatch
> BenchmarkProvideFor/NoMatch-12           7663920               153.8 ns/op            48 B/op          1 allocs/op
> BenchmarkProvideFor/OneMatch
> BenchmarkProvideFor/OneMatch-12          7447911               174.2 ns/op            48 B/op          1 allocs/op
> BenchmarkProvideFor/TwoMatches
> BenchmarkProvideFor/TwoMatches-12        2991748               409.7 ns/op           144 B/op          2 allocs/op
> BenchmarkProvideFor/IsData
> BenchmarkProvideFor/IsData-12            5941276               215.1 ns/op            96 B/op          2 allocs/op
> ```
> 
> After:
> 
> ```
> BenchmarkProvideFor
> BenchmarkProvideFor/Simple
> BenchmarkProvideFor/Simple-12           64254204                18.34 ns/op            0 B/op          0 allocs/op
> BenchmarkProvideFor/NoMatch
> BenchmarkProvideFor/NoMatch-12          30855906                38.58 ns/op            0 B/op          0 allocs/op
> BenchmarkProvideFor/OneMatch
> BenchmarkProvideFor/OneMatch-12          7462225               170.0 ns/op            48 B/op          1 allocs/op
> BenchmarkProvideFor/TwoMatches
> BenchmarkProvideFor/TwoMatches-12        3076194               407.0 ns/op           144 B/op          2 allocs/op
> BenchmarkProvideFor/IsData
> BenchmarkProvideFor/IsData-12           12269392               125.6 ns/op            48 B/op          1 allocs/op
> ```
> 
> i.e. this makes it faster in three of the five cases, mostly losing when we actually have a match. We might be able to get them down a little more with some more :thonk:

I love the tooling that go provides out of the box. No need to guess what's happening under the hood!