@peterebden 2021-06-17
> Looks good! I like these benchmark tests. Something we could use as metrics too

Yup agreed - although this one is pretty low-level, but it'd be nice to know it's not getting regressed!